set trigger(n) to n=10 in example
In the example we set the timer to 60 seconds. The trigger runs n
sequential consolidations. Each consolidation as a cross-subnet
call takes about 4-5 seconds normally. So we should be able to
complete 10 in a minute. If there is a large backlog then it is
better to process as much as possible of it.