Rewrite withdraw.test mit simple state tester
Using the variable tester and the set() function can actually hide some bugs.
For example, a set value be accidentally not read at all or read twice and we
would not notice. With the stage tester this will always be caught because it
will lead to a pop-from-empty-queue error.

This rewrite removed a couple of unneeded set() calls.