move write operations on entry outside consolidate()
Now, the consolidate() function only reads from entry. The write operations
entry.lock() and entry.setDeposit() are now bundled in trigger(). While the
entry is locked no other code can write to it. Therefore, it is ok to call
entry.setDeposit(0) after the consolidation transfer happens. It does not
necessarily have to be called before.